fix: 환불·취소 Outbox 이벤트 누락 보완 및 결제 도메인 무결성 강화(PDSH-17)

+++ payment/src/main/java/com/live_commerce/payment/infrastructure/client/dto/KakaoPayApproveDto.java
@@ -14,11 +14,11 @@ public record KakaoPayApproveDto(
 ) {
 
 	public record Amount(
-		@JsonProperty("total") int total,
-		@JsonProperty("tax_free") int taxFree,
-		@JsonProperty("vat") int vat,
-		@JsonProperty("point") int point,
-		@JsonProperty("discount") int discount
+		@JsonProperty("total") long total,
+		@JsonProperty("tax_free") long taxFree,
+		@JsonProperty("vat") long vat,
+		@JsonProperty("point") long point,
+		@JsonProperty("discount") long discount
 	) {}
 }
